Refactor image byte retrieval logic
