+++ src/ShapeCrawler/Drawing/SlidePictureImage.cs
@@ -18,7 +18,7 @@ internal sealed class SlidePictureImage : IImage
         this.aBlip = aBlip;
         this.sdkImagePart = (ImagePart)this.sdkTypedOpenXmlPart.GetPartById(aBlip.Embed!.Value!);
     }
-    
+
     public string MIME => this.sdkImagePart.ContentType;
 
     public string Name => Path.GetFileName(this.sdkImagePart.Uri.ToString());
@@ -52,14 +52,6 @@ internal sealed class SlidePictureImage : IImage
         byte[] sourceBytes = File.ReadAllBytes(file);
         this.Update(sourceBytes);
     }
-    
-    public byte[] AsByteArray()
-    {
-        var stream = this.sdkImagePart.GetStream();
-        var bytes = new WrappedStream(stream).AsBytes();
-        
-        stream.Close();
 
-        return bytes;
-    }
+    public byte[] AsByteArray() => new WrappedImagePart(this.sdkImagePart).AsBytes();
 }
