--- src/ShapeCrawler/Drawing/WrappedStream.cs
@@ -1,26 +1,0 @@
-﻿using System.IO;
-
-namespace ShapeCrawler.Drawing;
-
-internal readonly ref struct WrappedStream
-{
-    private readonly Stream stream;
-
-    internal WrappedStream(Stream stream)
-    {
-        this.stream = stream;
-    }
-
-    internal byte[] AsBytes()
-    {
-        var mStream = new MemoryStream();
-        var buffer = new byte[1024];
-        int read;
-        while ((read = this.stream.Read(buffer, 0, buffer.Length)) > 0)
-        {
-            mStream.Write(buffer, 0, read);
-        }
-        
-        return mStream.ToArray();
-    }
-}
